estilado favorite e historial de documentos, header pasado a componente
